+++ zarrs_filesystem/CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs_filesystem-v0.3.11...HEAD)
+## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs_filesystem-v0.3.12...HEAD)
+
+## [0.3.12](https://github.com/zarrs/zarrs/releases/tag/zarrs_filesystem-v0.3.12) - 2026-07-23
+
+### Added
+- Implement `AtomicRenameStorageTraits` for `FilesystemStore`
+
+### Changed
+- Bump `zarrs_storage` to 0.4.5
+- Increase the minimum supported Rust version to 1.87
 
 ## [0.3.11](https://github.com/zarrs/zarrs/releases/tag/zarrs_filesystem-v0.3.11) - 2026-07-21