+++ zarrs_storage/CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs_storage-v0.4.4...HEAD)
+## [Unreleased](https://github.com/zarrs/zarrs/compare/zarrs_storage-v0.4.5...HEAD)
+
+## [0.4.5](https://github.com/zarrs/zarrs/releases/tag/zarrs_storage-v0.4.5) - 2026-07-23
 
 ### Added
+- Add `AtomicWriteStorageAdapter` and `AtomicRenameStorageTraits`
 - Add `AsyncMemoryStore`
 
 ## [0.4.4](https://github.com/zarrs/zarrs/releases/tag/zarrs_storage-v0.4.4) - 2026-07-05